+++ libs/model/src/services/openai/generateTokenIdea.ts
@@ -1,15 +1,15 @@
-import { blobStorage } from '@hicommonwealth/core';
-import fetch from 'node-fetch';
 import { OpenAI } from 'openai';
 import {
   ChatCompletionMessage,
   ChatCompletionUserMessageParam,
 } from 'openai/resources/index.mjs';
-import { v4 as uuidv4 } from 'uuid';
 import { config } from '../../config';
 import { models } from '../../database';
 import { LaunchpadTokenInstance } from '../../models/token';
-import { compressServerImage } from '../../utils/imageCompression';
+import {
+  generateImage,
+  ImageGenerationErrors,
+} from '../../utils/generateImage';
 
 type GenerateTokenIdeaProps = {
   ideaPrompt?: string;
@@ -42,8 +42,6 @@ const TokenErrors = {
   RequestFailed: 'failed to generate complete token idea',
   IdeaPromptVoilatesSecurityPolicy:
     'provided `ideaPrompt` voilates content security policy',
-  ImageGenerationFailure: 'failed to generate image for token idea',
-  RunwareNotConfigured: 'Runware API key not configured',
 };
 
 // we have to maintain this to have "conversational" context with OpenAI
@@ -66,53 +64,6 @@ const chatWithOpenAI = async (prompt = '', openai: OpenAI) => {
   );
   console.log('Received response from OpenAI:', result);
   return result;
-};
-
-const generateImageWithRunware = async (prompt: string) => {
-  console.log('Generating token image with Runware model: runware:100@1');
-  const response = await fetch('https://api.runware.ai/v1/images/generations', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      Authorization: `Bearer ${config.IMAGE_GENERATION.RUNWARE_API_KEY}`,
-    },
-    body: JSON.stringify([
-      {
-        taskType: 'imageInference',
-        taskUUID: uuidv4(),
-        positivePrompt: prompt,
-        width: 256,
-        height: 256,
-        model: 'runware:100@1',
-        numberResults: 1,
-        outputType: 'URL',
-        outputFormat: 'PNG',
-      },
-    ]),
-  });
-
-  if (!response.ok) {
-    const error = new Error(`Runware API error: ${response.statusText}`);
-    console.error('Runware API error:', error);
-    throw error;
-  }
-
-  const data = await response.json();
-  console.log('Successfully generated token image with Runware', data);
-  return data.data[0].imageURL;
-};
-
-const generateImageWithOpenAI = async (prompt: string, openai: OpenAI) => {
-  console.log('Generating token image with OpenAI model: dall-e-3');
-  const imageResponse = await openai.images.generate({
-    prompt,
-    size: '1024x1024',
-    model: 'dall-e-3',
-    n: 1,
-    response_format: 'url',
-  });
-  console.log('Successfully generated token image with OpenAI');
-  return imageResponse.data[0].url || '';
 };
 
 const generateTokenIdea = async function* ({
@@ -172,40 +123,13 @@ const generateTokenIdea = async function* ({
     yield `data: ${tokenDescription}\n\n`;
 
     // generate image url and send the generated url to the client (to save time on s3 upload)
-    let imageUrl: string;
-    console.log(
-      `Using image generation service: ${config.IMAGE_GENERATION.FLAG_USE_RUNWARE ? 'Runware' : 'OpenAI'}`,
+    const imageUrl = await generateImage(
+      TOKEN_AI_PROMPTS_CONFIG.image(tokenName, tokenSymbol),
+      openai,
     );
 
-    if (config.IMAGE_GENERATION.FLAG_USE_RUNWARE) {
-      if (!config.IMAGE_GENERATION.RUNWARE_API_KEY) {
-        throw new Error(TokenErrors.RunwareNotConfigured);
-      }
-      imageUrl = await generateImageWithRunware(
-        TOKEN_AI_PROMPTS_CONFIG.image(tokenName, tokenSymbol),
-      );
-    } else {
-      imageUrl = await generateImageWithOpenAI(
-        TOKEN_AI_PROMPTS_CONFIG.image(tokenName, tokenSymbol),
-        openai,
-      );
-    }
-
-    // upload image to s3 and then send finalized imageURL
-    console.log('Uploading generated image to S3');
-    const resp = await fetch(imageUrl);
-    const buffer = await resp.buffer();
-    const compressedBuffer = await compressServerImage(buffer);
-    const { url } = await blobStorage().upload({
-      key: `${uuidv4()}.png`,
-      bucket: 'assets',
-      content: compressedBuffer,
-      contentType: 'image/png',
-    });
-    console.log('Successfully uploaded image to S3:', url);
-
     yield 'event: imageURL\n';
-    yield `data: ${url}\n\n`;
+    yield `data: ${imageUrl}\n\n`;
   } catch (e) {
     console.error('Error in generateTokenIdea:', e);
     let error = TokenErrors.RequestFailed;
@@ -217,7 +141,7 @@ const generateTokenIdea = async function* ({
       if (ideaPrompt && !tokenName) {
         error = TokenErrors.IdeaPromptVoilatesSecurityPolicy;
       } else {
-        error = TokenErrors.ImageGenerationFailure;
+        error = ImageGenerationErrors.ImageGenerationFailure;
       }
     }
